moved footer to molecules and changed ipsum
